+++ app/src/main/java/kz/post/jcourier/data/model/task/Task.kt
@@ -7,12 +7,12 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 @Keep
 data class Task(
-    var id: Int? = null,
-    var userId: Int? = null,
+    var id: Long? = null,
+    var userId: Long? = null,
     var status: TaskStatus? = null,
-    var serviceId: Int? = null,
-    var orderId: Int? = null,
-    var routeId: Int? = null,
+    var serviceId: Long? = null,
+    var orderId: Long? = null,
+    var routeId: Long? = null,
     var addressTo: AddressTo? = AddressTo(),
     var addressFrom: AddressFrom? = AddressFrom(),
     var contactTo: ContactTo? = ContactTo(),
